Fix warnings caused by header order

+++ dayz2/SpriteBatch.cpp
@@ -1,5 +1,4 @@
 #include "SpriteBatch.h"
-#include <GLFW\glfw3.h>
 
 CSpriteBatch::CSpriteBatch(int maxSprites) : m_bufferData(ELEMENTS_PER_VERTEX * 6 * maxSprites)
 {
@@ -64,7 +63,7 @@ void CSpriteBatch::setColor(float red, float green, float blue, float alpha)
 	m_a = alpha;
 }
 
-void CSpriteBatch::draw(float x, float y, float width, float height, float u1, float v1, float u2, float v2, float rotation)
+void CSpriteBatch::draw(float x, float y, int width, int height, float u1, float v1, float u2, float v2, float rotation)
 {
 	m_vertices += 6;
